% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Сообщение при подключении пакета}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Путь к библиотеке (передаётся R).}

\item{pkgname}{Имя пакета (передаётся R).}
}
\value{
NULL, невидимо.
}
\description{
В интерактивной сессии выводит версию пакета и ссылку на репозиторий.
}
\keyword{internal}
