% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_yaf_token}
\alias{get_yaf_token}
\title{Чтение сохранённого токена}
\usage{
get_yaf_token(login)
}
\arguments{
\item{login}{Логин в Яндексе.}
}
\value{
Токен (строка).
}
\description{
Читает токен, сохранённый yaf_get_token(). Если файла нет, в
интерактивной сессии запускает yaf_get_token(), в неинтерактивной
(Rscript, cron) останавливается с ошибкой.
}
\keyword{internal}
