% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Значение по умолчанию для NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Значение.}

\item{y}{Что вернуть, если x равен NULL.}
}
\value{
x, если он не NULL, иначе y.
}
\description{
Значение по умолчанию для NULL
}
\keyword{internal}
