% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{yaf_all_campaign_ids}
\alias{yaf_all_campaign_ids}
\title{Все Id кампаний аккаунта}
\usage{
yaf_all_campaign_ids(login)
}
\arguments{
\item{login}{Логин в Яндексе.}
}
\value{
Числовой вектор Id кампаний.
}
\description{
Все Id кампаний аккаунта
}
\keyword{internal}
