% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{yaf_api_request}
\alias{yaf_api_request}
\title{Один запрос к API с ретраями}
\usage{
yaf_api_request(
  login,
  service,
  body,
  token,
  max_retries = 5,
  retry_codes = c(52, 506, 1000, 1001, 1002)
)
}
\description{
Один запрос к API с ретраями
}
\keyword{internal}
