% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{yaf_changes_ids}
\alias{yaf_changes_ids}
\title{Разбор блока id из ответа changes.check}
\usage{
yaf_changes_ids(x)
}
\arguments{
\item{x}{Блок Modified, NotFound или Unprocessed из ответа API.}
}
\value{
data.frame с колонками Type ("Campaign", "AdGroup", "Ad") и Id.
}
\description{
Превращает объект вида list(CampaignIds = list(...), AdGroupIds = ...,
AdIds = ...) в длинную таблицу. Поле CampaignsStat пропускается.
}
\keyword{internal}
