% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{yaf_find_goal_id}
\alias{yaf_find_goal_id}
\title{Id цели из параметров стратегии}
\usage{
yaf_find_goal_id(x)
}
\arguments{
\item{x}{Стратегия площадки (BiddingStrategy$Search / $Network)
или объект пакетной стратегии.}
}
\value{
Число или NA, если у стратегии нет цели.
}
\description{
GoalId лежит в блоке параметров конкретной стратегии
(AverageCpa$GoalId, PayForConversion$GoalId и т. д.).
}
\keyword{internal}
