% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{yaf_format_timestamp}
\alias{yaf_format_timestamp}
\title{Приведение момента времени к формату сервиса changes}
\usage{
yaf_format_timestamp(timestamp)
}
\arguments{
\item{timestamp}{Строка в формате "YYYY-MM-DDThh:mm:ssZ" или POSIXct.}
}
\value{
Строка в формате "YYYY-MM-DDThh:mm:ssZ".
}
\description{
Сервис changes принимает время в UTC в формате ISO 8601:
"YYYY-MM-DDThh:mm:ssZ".
}
\keyword{internal}
