% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{yaf_get_regions}
\alias{yaf_get_regions}
\title{Получение справочника регионов Яндекс Директа}
\usage{
yaf_get_regions(login)
}
\arguments{
\item{login}{Логин в Яндексе.}
}
\value{
data.frame с колонками region_id, parent_id, region_name,
region_type. При ошибке API — NULL.
}
\description{
Возвращает дерево регионов (ID, название, родительский ID, тип).
https://yandex.ru/dev/direct/doc/ru/dictionaries/get
}
\examples{
\dontrun{
regions <- yaf_get_regions("my_login")
}
}
