% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yaf_get_token}
\alias{yaf_get_token}
\title{Получение и сохранение токена Яндекс Директа}
\usage{
yaf_get_token(login)
}
\arguments{
\item{login}{Логин в Яндексе.}
}
\value{
Токен (строка), невидимо.
}
\description{
Открывает в браузере страницу авторизации Яндекса, просит вставить
полученный OAuth-токен в консоль и сохраняет его в файл \verb{<login>.rds}
в папке \code{tools::R_user_dir("yaf", "config")}. Остальные функции пакета
читают токен из этого файла. Работает только в интерактивной сессии.
}
\examples{
\dontrun{
yaf_get_token("my_login")
}
}
