% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{yaf_parse_units}
\alias{yaf_parse_units}
\title{Разбор заголовка Units: "потрачено/осталось/лимит"}
\usage{
yaf_parse_units(header)
}
\description{
Разбор заголовка Units: "потрачено/осталось/лимит"
}
\keyword{internal}
