% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{yaf_priority_goal_ids}
\alias{yaf_priority_goal_ids}
\title{Id целей из PriorityGoals}
\usage{
yaf_priority_goal_ids(x)
}
\arguments{
\item{x}{Объект с полем PriorityGoals = list(Items = list(list(GoalId, ...))).}
}
\value{
Числовой вектор id целей.
}
\description{
Id целей из PriorityGoals
}
\keyword{internal}
