% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{yaf_split_keyword}
\alias{yaf_split_keyword}
\title{Разделение фразы на саму фразу и минус-слова}
\usage{
yaf_split_keyword(keyword)
}
\description{
"---autotargeting" (автотаргетинг) минус-словом не считается.
}
\keyword{internal}
