% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_yaf_token}
\alias{get_yaf_token}
\title{Internal function to retrieve token}
\usage{
get_yaf_token(login)
}
\arguments{
\item{login}{Character. Yandex login.}
}
\description{
Internal function to retrieve token
}
\keyword{internal}
