% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{yaf_api_error}
\alias{yaf_api_error}
\title{Ошибка API с кодом}
\usage{
yaf_api_error(
  message,
  code = NA_integer_,
  service = NA_character_,
  request_id = NA_character_
)
}
\description{
Ошибка API с кодом
}
\keyword{internal}
