% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yaf_fields_info}
\alias{yaf_fields_info}
\title{Справочник полей API Яндекс Директа}
\format{
A data frame with columns:
\describe{
\item{FieldNames}{Название поля для аргумента fields}
}
}
\usage{
yaf_fields_info
}
\description{
Таблица содержит список доступных полей, их типы и совместимость с отчетами.
}
\keyword{datasets}
