% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adgroups.R
\name{yaf_get_adgroups}
\alias{yaf_get_adgroups}
\title{Получение списка групп объявлений из Яндекс Директа}
\usage{
yaf_get_adgroups(login, campaign_ids = NULL)
}
\arguments{
\item{login}{Логин в Яндексе.}

\item{campaign_ids}{Вектор ID кампаний (необязательно).}

\item{adgroup_ids}{Вектор ID конкретных групп (необязательно).}
}
\description{
Получение списка групп объявлений из Яндекс Директа
}
