% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\name{yaf_get_ads}
\alias{yaf_get_ads}
\title{Получение объявлений из Яндекс Директа}
\usage{
yaf_get_ads(login, campaign_ids = NULL, adgroup_ids = NULL, ad_ids = NULL)
}
\arguments{
\item{login}{Логин в Яндексе.}

\item{campaign_ids}{Вектор ID кампаний (необязательно).}

\item{adgroup_ids}{Вектор ID групп (необязательно).}

\item{ad_ids}{Вектор ID конкретных объявлений (необязательно).}
}
\description{
Получение объявлений из Яндекс Директа
}
