% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidmodifiers.R
\name{yaf_get_bid_modifiers}
\alias{yaf_get_bid_modifiers}
\title{Получение корректировок ставок из Яндекс Директа}
\usage{
yaf_get_bid_modifiers(login, campaign_ids = NULL, adgroup_ids = NULL)
}
\arguments{
\item{login}{Логин в Яндексе.}

\item{campaign_ids}{Вектор ID кампаний (необязательно).}

\item{adgroup_ids}{Вектор ID групп (необязательно).}
}
\description{
Получение корректировок ставок из Яндекс Директа
}
