% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{yaf_get_campaigns}
\alias{yaf_get_campaigns}
\title{Получение списка рекламных кампаний}
\usage{
yaf_get_campaigns(login)
}
\arguments{
\item{login}{Character. Логин в Яндексе.}
}
\value{
A data frame с ID кампаний, их названиями, статусами и типами.
}
\description{
Функция возвращает таблицу со списком всех кампаний рекламодателя.
}
\examples{
\dontrun{
my_campaigns <- yaf_get_campaigns("my_login")
}
}
