% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{yaf_get_report}
\alias{yaf_get_report}
\title{Получение статистики из Яндекс Директа}
\usage{
yaf_get_report(login, date_from, date_to, fields, goals = NULL)
}
\arguments{
\item{login}{Character. Логин в Яндексе.}

\item{date_from}{Character. Дата начала (ГГГГ-ММ-ДД).}

\item{date_to}{Character. Дата окончания (ГГГГ-ММ-ДД).}

\item{fields}{Character vector. Список полей для выгрузки (обязательный).}

\item{goals}{Numeric vector. Список ID целей Метрики (необязательно).}
}
\description{
Получение статистики из Яндекс Директа
}
