% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{yaf_get_report}
\alias{yaf_get_report}
\title{Получение статистики из Яндекс Директа}
\usage{
yaf_get_report(
  login,
  date_from = Sys.Date() - 7,
  date_to = Sys.Date() - 1,
  date_range_type = NULL,
  fields = c("Date", "Impressions", "Clicks"),
  goals = NULL,
  atribution = NULL,
  filter = NULL,
  search_query_report = FALSE,
  numeric_fields_as_numeric = TRUE,
  include_vat = TRUE,
  save_report = FALSE
)
}
\arguments{
\item{login}{Логин в Яндексе.}

\item{date_from}{Дата начала (ГГГГ-ММ-ДД).}

\item{date_to}{Дата окончания (ГГГГ-ММ-ДД).}

\item{date_range_type}{Пресет периода (YESTERDAY, TODAY, LAST_7_DAYS, LAST_30_DAYS, LAST_MONTH).}

\item{fields}{Список полей для выгрузки.}

\item{goals}{Список ID целей Метрики (необязательно).}

\item{atribution}{Модели атрибуции.}

\item{filter}{Строка для фильтрации данных.}

\item{search_query_report}{TRUE для отчета по поисковым запросам.}

\item{numeric_fields_as_numeric}{Не менять тип полей.}

\item{include_vat}{Включать ли НДС (YES/NO).}

\item{save_report}{Логическое значение. Если TRUE, сохраняет CSV в папку /reports.}
}
\description{
Получение статистики из Яндекс Директа
}
