% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yaf_get_token}
\alias{yaf_get_token}
\title{Get and save Yandex Direct Token to RDS}
\usage{
yaf_get_token(login)
}
\arguments{
\item{login}{Character. Your Yandex login.}
}
\description{
Opens a browser to get an OAuth token and saves it to a local RDS file for future use.
}
